Include Templates.pkgdef
A few things going on here.

The default VSIX project assumes you are going to deploy code and an actual extension.
If this is the scenario, you need to create a serperate pkgdef and reference that as shown, since the tooling will overwrite the
package def if you edit it - the VsixTemplate.pkgdef was being auto generated and the static one was not being picked up.

If you want to only include templates, you can remove the code and generation of the DLL and remove the VsPackage with Source=Project
This is what the sample at http://github.com/phenning/templateSamples does.